Refactor logging encoders to utilize shared utility functions for field extraction and mapping
